hackathon_aws_version/backend/requirements.txt
flask
flask-cors
requests
python-dotenv
openai
numpy
scikit-learn
